src/trollbox/requirements.txt
arrow==1.0.3
binaryornot==0.4.4
briefcase==0.3.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
Jinja2==2.11.3
jinja2-time==0.3.0
keyboard==0.13.5
MarkupSafe==1.1.1
mouse==0.7.1
poyo==0.5.0
pycparser==2.20
PySDL2==0.9.7
pysdl2-dll==2.0.14.post1
python-dateutil==2.8.1
python-slugify==4.0.1
pythonnet==2.5.2
requests==2.25.1
six==1.15.0
smmap==4.0.0
text-unidecode==1.3
toga==0.3.0.dev26
toga-core==0.3.0.dev26
toga-winforms==0.3.0.dev26
toml==0.10.2
travertino==0.1.3
urllib3==1.26.4
